Handle unexpected response payloads from Instaclustr more gracefully
Sometimes Instaclustr API may return partial data, for example one node in the
cluster may be missing data whereas other have it. This is a symptom of issues
at Instaclustr side - something that often, but not always, resolves itself. In
these cases warn that the payload received as a response could not be parsed and
exit with a warning.

Signed-off-by: Samuli Seppänen <samuli.seppanen@puppeteers.net>